Revamp type import in projects
